+++ src/runtime/engine/utils.ts
@@ -16,14 +16,6 @@ export const rand = (min: number, max: number): number => {
   return Math.random() * (max - min) + min
 }
 
-export const DIRECTIONS: Record<string, { x: number, y: number }> = {
-  none: { x: 0, y: 0 },
-  top: { x: 0, y: -1 },
-  bottom: { x: 0, y: 1 },
-  left: { x: -1, y: 0 },
-  right: { x: 1, y: 0 },
-}
-
 export const mergeConfig = (
   base: ParticleConfig,
   over: ParticleConfig,